+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - [ ] make it work embedded in HC tech blog
   - [x] show errors in the data format, to not throw errors when trying to draw the chart
   - [x] dont bubble key events from the editor, the techblog uses them to open the menu and other strange stuff
+  - [x] don't throw error when label is not a string